+++ packages/saas-ui-card/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @saas-ui/card
 
+## 1.0.0-rc.4
+
+### Patch Changes
+
+- 68c7b62: Card now has a isHoverable property to support hover styles.
+- Updated dependencies [870db1f]
+  - @saas-ui/button@1.0.0-rc.4
+
 ## 1.0.0-rc.3
 
 ### Patch Changes